Phase 7 hardening: canonical stripe webhook and portal submit warnings
