+++ docs/phase-7-final-hardening-summary.md
@@ -1,0 +1,54 @@
+# Final Hardening Summary — Test Readiness
+
+## Phases completed
+- Phase 1: authz/tenant isolation hardening audit confirmed and re-validated in touched callers
+- Phase 2: owner PIN security hardening carried forward with existing crypto coverage
+- Phase 3: billing canonicalization paths retained and finalized on canonical Stripe webhook source-of-truth path
+- Phase 4: approval canonicalization paths retained (customer approval webhook remains canonical)
+- Phase 5: booking integrity flow retained and hardened for degraded-path observability
+- Phase 6: canonical cleanup follow-through completed (legacy Stripe feature path now true compatibility alias)
+- Phase 7: final test-readiness hardening completed in this phase
+
+## Systems hardened
+- Stripe billing webhook source-of-truth canonicalized to `features/stripe/api/stripe/webhook/route.ts` and legacy feature endpoint reduced to import-only alias.
+- Portal request submit runtime hardening for non-fatal degraded paths, with explicit warnings returned to callers and structured warning logs.
+- Added high-value webhook smoke coverage focused on environment/config and request-shape assumptions that commonly block E2E cycles.
+
+## Canonical paths in effect
+- Stripe webhook canonical logic: `features/stripe/api/stripe/webhook/route.ts`
+- App Stripe webhook entrypoint: `/api/stripe/webhook`
+- Legacy Stripe app route compatibility alias: `/api/stripe/checkout/webhook` → canonical handler
+- Legacy Stripe feature compatibility alias: `features/stripe/api/stripe/checkout/webhook/route.ts` → canonical feature handler
+- Portal request submit canonical flow: `/api/portal/request/submit` with explicit degraded-path warnings
+
+## Files changed in this phase
+- app/api/portal/request/submit/route.ts
+- features/stripe/api/stripe/webhook/route.ts
+- features/stripe/api/stripe/checkout/webhook/route.ts
+- tests/stripe-webhook-hardening.test.ts
+- docs/phase-7-final-hardening-summary.md
+
+## Migrations added in this phase
+- None
+- manual SQL apply required: no (no new SQL migration in this phase)
+- regenerate Supabase types after apply: no-op for this phase (still required whenever future SQL is applied)
+
+## Validation run
+- `npx tsc --noEmit` → pass
+- `npx vitest run tests/owner-pin-crypto.test.ts tests/portal-intake.test.ts tests/stripe-webhook-hardening.test.ts tests/smoke.test.ts` → pass (9 tests)
+- `npx eslint app/api/portal/request/submit/route.ts features/stripe/api/stripe/webhook/route.ts features/stripe/api/stripe/checkout/webhook/route.ts tests/stripe-webhook-hardening.test.ts` → pass
+
+## Remaining known risks before full E2E
+- Portal submit intentionally preserves non-blocking behavior for downstream parts request creation; failures are now explicit via `warnings`, but full operational alerting/telemetry sink integration is still external to this phase.
+- Legacy compatibility endpoints remain by design and should only be removed after all external callers are verified migrated.
+
+## Manual test priorities for next 5 days
+1. Portal request start → submit happy path and replay/idempotency path under concurrent user retries.
+2. Portal submit degraded path behavior: validate `warnings` visibility when part-request RPC or intake-line persistence is intentionally faulted.
+3. Stripe webhook subscription checkout path (`checkout.session.completed` subscription mode) and downstream shop/profile billing state propagation.
+4. Stripe webhook one-time payment path (`checkout.session.completed` payment mode) including payment upsert idempotency.
+5. Customer approval webhook decisions with mixed approved/declined lines and repeated submissions (idempotent behavior).
+6. Owner PIN set/verify/clear across owner/admin/staff roles and wrong-shop actor attempts.
+
+## Launch-readiness status after hardening
+- Conditionally ready for aggressive E2E execution: critical touched flows now have clearer failure surfacing, canonical webhook source-of-truth alignment, and added smoke coverage for high-frequency env/config blockers.
